Templatize basic test suite
