+++ package.json
@@ -72,9 +72,11 @@
     "@polymer/paper-button": "^3.0.1",
     "@polymer/paper-radio-group": "^3.0.1",
     "@polymer/paper-slider": "^3.0.1",
+    "@types/chai": "^4.1.7",
+    "@types/mocha": "^5.2.5",
     "@types/pngjs": "^3.3.2",
-    "@types/puppeteer": "^1.11.1",
     "@types/prismjs": "^1.9.0",
+    "@types/puppeteer": "^1.11.1",
     "@types/three": "^0.93.12",
     "@webcomponents/webcomponentsjs": "~2.1.3",
     "chai": "^4.1.2",
@@ -89,8 +91,8 @@
     "mocha": "^5.2.0",
     "pngjs": "^3.3.3",
     "polymer-build": "^3.1.1",
-    "puppeteer": "^1.11.0",
     "prismjs": "^1.15.0",
+    "puppeteer": "^1.11.0",
     "resize-observer-polyfill": "^1.5.0",
     "rimraf": "^2.6.2",
     "rollup": "^0.66.0",
