+++ src/test/model-viewer-base-spec.js
@@ -16,6 +16,7 @@
 import ModelViewerElementBase, {$canvas, $renderer, $scene} from '../model-viewer-base.js';
 
 import {assetPath, timePasses, until, waitForEvent} from './helpers.js';
+import {BasicSpecTemplate} from './templates.js';
 
 const expect = chai.expect;
 
@@ -39,44 +40,7 @@ suite('ModelViewerElementBase', () => {
       customElements.define(tagName, ModelViewerElement);
     });
 
-    test('can be directly instantiated', () => {
-      const element = new ModelViewerElement();
-      expect(element).to.be.ok;
-    });
-
-    test('can be instantiated with document.createElement', () => {
-      const element = document.createElement(tagName);
-      expect(element).to.be.ok;
-    });
-
-    suite('compatibility', () => {
-      suite('when WebGL is not supported', () => {
-        let nativeGetContext;
-
-        setup(() => {
-          nativeGetContext = HTMLCanvasElement.prototype.getContext;
-          HTMLCanvasElement.prototype.getContext = function(type, ...args) {
-            if (/webgl/.test(type)) {
-              return null;
-            }
-            return nativeGetContext.call(this, type, ...args);
-          };
-        });
-
-        teardown(() => {
-          HTMLCanvasElement.prototype.getContext = nativeGetContext;
-        });
-
-        test(
-            'does not explode when created and appended to the document',
-            async () => {
-              const element = new ModelViewerElement();
-              document.body.appendChild(element);
-              await timePasses();
-              document.body.removeChild(element);
-            });
-      });
-    });
+    BasicSpecTemplate(() => ModelViewerElement, () => tagName);
 
     suite('with alt text', () => {
       let element;
